requirements.txt
pytest==8.3.4
allure-pytest==2.13.5
allure-python-commons==2.13.5
selenium==4.22.0
Faker==35.0.0
requests==2.32.3